requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
crawlee==1.0.0
apify-client==1.4.0
markdownify==0.11.6
google-generativeai==0.3.2
playwright==1.40.0 